Making the ActivityNameEntryBox not miss inputs when typing quickly.
VisiPlacer now knows not to temporarily disconnect a View when its parent doesn't change.
VisiPlacer doesn't yet have a good solution for cases when the parent does change,
so for the ActivityNameEntryBox we always assign the same grid as the parent of its textBox, to prevent the textBox from temporarily disappearing after typing one letter and sometimes missing the second letter typed.