Visually packing the ParticipationEntryView more efficiently
The activity name autocomplete suggestion doesn't need to be visible at the same time as the participation feedback.
If we enforce this, then we can make the two use the same space.

Also, the words "Activity Name" are probably unnecessary given that the title asks the user what they've been doing, and there's already a "?" button.

Also, the metric and completion status buttons don't need space allocated to them when they're not visible.

Also, we can ask the comment box to show as much text as will fit in it.