More making the engine less likely to repeat its suggestions
Previously, as the user entered more skips, the engine would lower the participation probability and would reduce the weight of some estimates.
That would cause GetCombinedValue to report a higher potential score for that activity due to it thinking that there wasn't much data.

Now the weight of most estimates should increase even when a suggestion is skipped, making repetition less likely