Restoring the ability for the ActivityCreationLayout to create multiple activity types
because this is faster and clearer when the user wants to create a Category, especially from within the ParticipationEntryView's activity name entry box.

Also moving the ActivityCreationLayout higher in the menu hierarchy so it's faster to find, now that it's back to being one layout.
Also adding more clarification that it can create both Categories and Problems, so a user that wants to create a Problem knows to select it.