Unifying the activity-browsing screens
to facilitate queries like "What are all of the open ToDos under activity X?" before starting an experiment.

and to reduce by 1 the number of button presses required to get to the search view, which most commonly used.

Also making the ActivityInheritancesView compute its layout more quickly (by not attempting to show multiple scroll
layouts at once, and instead requiring the user to choose a scroll layout to look at) to facilitate adding this new
button to the activity inheritances view.