SuggestionsView moving the Experiment and Help buttons to the bottom
If a user asks for an experiment and another suggestion, the experiment needs to be first.
This suggests putting the Experiment button before the Suggest button.

However, the amount of information on this screen is getting to be kind of large compared to what
a new user is probably willing to think about, so we're moving the Experiment button to the bottom
to make it clearer to new users that they probably want to push the Suggest button