Moving the participation summary and preference summary into the View Statistics menu
Even though they provide analyses that are easy to share with other people, users are probably more likely to
consider them analyses first and shareable second, so we're moving them into the same screen with the other analyses.

Also, this shortens/rewords many descriptions to make it faster to read them and easier to recognize their differences.