Fixing Make_CompletionEfficiencyMeasurement to conserve true total effectiveness rather than first rescaling by difficulty
Previously this function was rescaling times by their expected difficulties.
Now this function directly computes the total effectiveness and redistributes it between the participations based on their expected difficulties.

This makes the most difference when the difficulties of the two tasks are significantly different, because
it makes the more-difficult participation not have to sacrifice as much efficiency when giving efficiency to the other participation.