SuggestionsView feedback now using the same logic and format as the ParticipationsEntryView feedback
to provide more consistency. This way, users should be able to feel that they can almost rely on the description
in the suggestion as a promise of what feedback they can expect when they complete the task.

This should make things more predictable and be more motivating.

It should also be more informative, because the happiness value appears as a number in the response text.